Revert version bump
While the Gravatar pod version 1.0.1-alpha has indeed been pushed to CocoaPods, this was a dummy version with no code change, only used to test the publish automation. Besides, GravatarUI pod version 1.0.1-alpha was never pushed alongside Gravatar pod 1.0.1-alpha.

We could consider deleting Gravatar 1.0.1-alpha from CocoaPods and remove the git tag, but it's usually considered bad practice to delete a version that has already been pushed; so it's not the end of the world to keep it in CocoaPods, but doesn't make sense to keep the `version.rb` with that dummy version though.