Remove `@testable import`s to enable Release builds
I am don't have the background on the rationale that went into deciding
`@testable import` over declaring the types that the annotation allowed
to access as `public`.

However, `@testable import` works under the assumption that a target is
built with testability enabled, which is not and should not be the case
in the Release configuration.

Removing the annotation required declaring the various types used by the
demo apps to be declared as `public`.

I'll leave it up to the team to decide whether my changes were
appropriate and to make any necessary adjustment. I did only what was
necessary to get the Release to compile without diving into the API
design and the effect that the updated access control would have.
Presumably, if a demo app needed those types, then so would a real world
app, but it's hard for me to tell as an infrastructure engineer jumping
in the codebase to implement Prototype Builds.